requirements.txt
bentoml==1.1.2
tensorflow==2.11.0
tensorflow_hub==0.12.0
google-cloud-storage==2.10.0
opencv-python-headless==4.8.0.76
# grpcio==1.57.0
